Refactored components 🏗️
